sudokusolver/requirements.txt
pygame>=2.0
Pillow>=9.0
anthropic>=0.20
